PIPELINE-4259: Rename pipelines/detect package to pipelines/raw_gaps
Mirrors the CLI subcommand naming: pipelines/detect/ -> pipelines/raw_gaps/,
DetectGapsConfig -> RawGapsConfig, DetectGapsLinearDagFactory ->
RawGapsLinearDagFactory. The DetectGaps Beam PTransform (transforms/detect_gaps.py)
is left unchanged, as it names a processing step rather than the command.
Updates cross-references from pipelines/events/table_config.py and docs.
