+++ README.md
@@ -57,7 +57,7 @@ Time gap detector for vessel position reporting systems (**[AIS]**, **[VMS]**, e
 
 [gaps.py]: src/pipe_gaps/queries/gaps.py
 [messages.py]: src/pipe_gaps/queries/messages.py
-[detect_gaps.py]: src/pipe_gaps/pipelines/detect/transforms/detect_gaps.py
+[detect_gaps.py]: src/pipe_gaps/pipelines/raw_gaps/transforms/detect_gaps.py
 [gap_detector.py]: src/pipe_gaps/core/gap_detector.py
 [assets/schemas/]: src/pipe_gaps/assets/schemas
 [schemas/gaps.json]: src/pipe_gaps/assets/schemas/gaps.json
@@ -95,8 +95,8 @@ Time gap detector for vessel position reporting systems (**[AIS]**, **[VMS]**, e
 
 <div align="justify">
 
-Not all vessel position reports generated by onboard reporting systems (such as **AIS** or **VMS**)
-are successfully recorded or received.
+Not all vessel position reports generated by onboard reporting systems
+(such as **AIS** or **VMS**) are successfully recorded or received.
 For **AIS**, this may occur due to signal interference in crowded waters,
 reception limitations,
 and variability in terrestrial or satellite coverage,
@@ -473,13 +473,13 @@ This is an example of a JSON config file for `raw-gaps` command:
 This is an example of a YAML config file for the publication `gap-events` command:
 ```yaml
 # Query parameters:
-bq_in_gaps: "gfw-int-pipe-v3.pipe_ais_v3_internal.raw_gaps_last_versions"
+bq_in_raw_gaps: "gfw-int-pipe-v3.pipe_ais_v3_internal.raw_gaps_last_versions"
 bq_in_segment_info: "global-fishing-watch.pipe_ais_v3_published.segment_info"
 bq_in_regions: "global-fishing-watch.pipe_regions_layers.event_regions"
 bq_in_voyages: "global-fishing-watch.pipe_ais_v3_published.voyages_c4"
 bq_in_port_visits: "global-fishing-watch.pipe_ais_v3_published.product_events_port_visit_v2"
 bq_in_all_vessels_byyear: "global-fishing-watch.pipe_ais_v3_published.product_vessel_info_summary"
-bq_output: "world-fishing-827.scratch_tomas_ttl30d.product_events_ais_gap"
+bq_out_gap_events: "world-fishing-827.scratch_tomas_ttl30d.product_events_ais_gap"
 date_range: ["2025-11-01", "2025-11-02"]
 
 # General parameters:
