+++ config/sample-gap-events.yaml
@@ -1,12 +1,12 @@
 # Query parameters:
-bq_in_gaps: "world-fishing-827.scratch_tomas_ttl30d.raw_gaps_test_view_last_versions"
+bq_in_raw_gaps: "world-fishing-827.scratch_tomas_ttl30d.raw_gaps_test_view_last_versions"
 bq_in_segment_info: "global-fishing-watch.pipe_ais_v3_published.segment_info"
 bq_in_segs_activity: "global-fishing-watch.pipe_ais_v3_published.segs_activity"
 bq_in_regions: "global-fishing-watch.pipe_regions_layers.event_regions"
 bq_in_voyages: "global-fishing-watch.pipe_ais_v3_published.voyages_c4"
 bq_in_port_visits: "global-fishing-watch.pipe_ais_v3_published.product_events_port_visit_v2"
 bq_in_vessels_byyear: "global-fishing-watch.pipe_ais_v3_published.product_vessel_info_summary"
-bq_output: "world-fishing-827.scratch_tomas_ttl30d.product_events_ais_gap"
+bq_out_gap_events: "world-fishing-827.scratch_tomas_ttl30d.product_events_ais_gap"
 date_range: ["2025-11-01", "2025-11-02"]
 
 # General parameters:
