+++ src/pipe_gaps/cli/commands/gap_events.py
@@ -11,7 +11,7 @@ DESCRIPTION = """\
 Enriches gaps data and creates publication events.
 """
 
-HELP_BQ_IN_GAPS = "BigQuery table with gaps."
+HELP_BQ_IN_RAW_GAPS = "BigQuery table with raw gaps."
 HELP_BQ_IN_SEGMENT_INFO = "BigQuery table with segments information."
 HELP_BQ_IN_SEGS_ACTIVITY = "BigQuery table with research aggregated segments data."
 HELP_BQ_IN_VOYAGES = "BigQuery table with voyages."
@@ -21,7 +21,7 @@ HELP_BQ_IN_VESSELS_BYYEAR = "BigQuery table with vessels by year."
 HELP_BQ_IN_VESSELS_BYYEAR_FIELD_PREFIX = "Field prefix for fields in bq-in-vessels-by-year."
 HELP_LABELS = "Labels to audit costs over the queries."
 
-HELP_BQ_OUTPUT = "BigQuery table in which to store the gap events."
+HELP_BQ_OUT_GAP_EVENTS = "BigQuery table in which to store the gap events."
 
 HELP_MOCK_BQ_CLIENTS = "If passed, mocks the BQ clients [Useful for development]."
 HELP_DATE_RANGE = "Create gap events for this date range, e.g., «2024-01-01,2024-01-02»."
@@ -44,7 +44,7 @@ class GapEvents(Command):
             Option("--date-range", type=date_range, help=HELP_DATE_RANGE),
             Option("--project", type=str, help=HELP_BQ_PROJECT),
             Option("--dry-run", type=bool, help=HELP_DRY_RUN),
-            Option("--bq-in-gaps", type=str, help=HELP_BQ_IN_GAPS),
+            Option("--bq-in-raw-gaps", type=str, help=HELP_BQ_IN_RAW_GAPS),
             Option("--bq-in-segment-info", type=str, help=HELP_BQ_IN_SEGMENT_INFO),
             Option("--bq-in-segs-activity", type=str, help=HELP_BQ_IN_SEGS_ACTIVITY),
             Option("--bq-in-voyages", type=str, help=HELP_BQ_IN_VOYAGES),
@@ -53,7 +53,7 @@ class GapEvents(Command):
             Option("--bq-in-vessels-byyear", type=str, help=HELP_BQ_IN_VESSELS_BYYEAR),
             Option("--bq-in-vessels-byyear-field-prefix", type=str,
                    help=HELP_BQ_IN_VESSELS_BYYEAR_FIELD_PREFIX, default=""),
-            Option("--bq-output", type=str, help=HELP_BQ_OUTPUT),
+            Option("--bq-out-gap-events", type=str, help=HELP_BQ_OUT_GAP_EVENTS),
             Option("--mock-bq-clients", type=bool, help=HELP_MOCK_BQ_CLIENTS),
             Option("--labels", type=str, nargs="*", action=NestedKeyValueAction, help=HELP_LABELS),
         ]
