+++ src/pipe_gaps/pipelines/events/config.py
@@ -8,7 +8,7 @@ from gfw.common.config import PipelineConfig
 
 @dataclass(frozen=True, kw_only=True)
 class GapEventsConfig(PipelineConfig):
-    bq_in_gaps: str
+    bq_in_raw_gaps: str
     bq_in_segment_info: str
     bq_in_segs_activity: str
     bq_in_voyages: str
@@ -16,7 +16,7 @@ class GapEventsConfig(PipelineConfig):
     bq_in_regions: str
     bq_in_vessels_byyear: str
     bq_in_vessels_byyear_field_prefix: str
-    bq_output: str
+    bq_out_gap_events: str
     project: str
     labels: dict = field(default_factory=dict)
     dry_run: bool = False
