+++ src/pipe_gaps/pipelines/raw_gaps/fns/process_boundaries.py
@@ -9,7 +9,7 @@ from pipe_gaps.core import GapDetector
 from pipe_gaps.common.key import Key
 from pipe_gaps.common.beam.side_inputs import SideInputs
 from pipe_gaps.common.sorting import timestamp_msgid_key
-from pipe_gaps.pipelines.detect.fns.extract_group_boundary import Boundary
+from pipe_gaps.pipelines.raw_gaps.fns.extract_group_boundary import Boundary
 
 logger = logging.getLogger(__name__)
