+++ tests/pipelines/events/test_main.py
@@ -8,7 +8,7 @@ from pipe_gaps.pipelines.events import main
 def basic_config_kwargs():
     return {
         "date_range": ("2024-01-01", "2024-01-02"),
-        "bq_in_gaps": "project.dataset.table",
+        "bq_in_raw_gaps": "project.dataset.table",
         "bq_in_segment_info": "project.dataset.table",
         "bq_in_segs_activity": "project.dataset.table",
         "bq_in_regions": "project.dataset.table",
@@ -16,7 +16,7 @@ def basic_config_kwargs():
         "bq_in_port_visits": "project.dataset.table",
         "bq_in_vessels_byyear": "project.dataset.table",
         "bq_in_vessels_byyear_field_prefix": "ais_",
-        "bq_output": "project.dataset.table",
+        "bq_out_gap_events": "project.dataset.table",
         "unknown_unparsed_args": [],
         "project": "test-project",
         "mock_bq_clients": True,
