+++ tests/pipelines/raw_gaps/fns/test_process_boundaries.py
@@ -3,8 +3,8 @@ from datetime import date, timedelta
 
 from pipe_gaps.core import GapDetector
 from pipe_gaps.common.key import Key
-from pipe_gaps.pipelines.detect.fns.extract_group_boundary import Boundary
-from pipe_gaps.pipelines.detect.fns.process_boundaries import (
+from pipe_gaps.pipelines.raw_gaps.fns.extract_group_boundary import Boundary
+from pipe_gaps.pipelines.raw_gaps.fns.process_boundaries import (
     Boundaries,
     ProcessBoundaries,
 )
