+++ tests/pipelines/raw_gaps/test_config.py
@@ -1,4 +1,4 @@
-from pipe_gaps.pipelines.detect.table_config import GapsTableConfig
+from pipe_gaps.pipelines.raw_gaps.table_config import GapsTableConfig
 
 
 def test_gaps_table_config_property(base_config):
